rfct: Move password generation to GeneratorFragment
