+++ app/src/main/kotlin/com/cyb3rko/pazzword/fragments/GeneratorFragment.kt
@@ -17,6 +17,7 @@
 package com.cyb3rko.pazzword.fragments
 
 import android.content.Context
+import android.os.Build
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
@@ -28,12 +29,30 @@ import com.cyb3rko.pazzword.R
 import com.cyb3rko.pazzword.databinding.FragmentGeneratorBinding
 import com.cyb3rko.pazzword.showClipboardToast
 import com.cyb3rko.pazzword.storeToClipboard
-import com.cyb3rko.pazzword.generatorPassword
-import com.cyb3rko.pazzword.PasswordTypes
 import com.google.android.material.slider.Slider
 import me.gosimple.nbvcxz.resources.Generator
+import java.security.SecureRandom
 
 class GeneratorFragment : Fragment() {
+    companion object {
+        const val TYPE_PASSPHRASE = 1
+        const val TYPE_PASSWORD = 2
+
+        const val alpha = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ"
+        const val alphaNumeric = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ1234567890"
+        const val alphaNumericSymbols = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ1234567890!@#$%^&*()"
+        const val numeric = "1234567890"
+        const val alphaSpace = " abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ"
+    }
+
+    enum class PasswordTypes {
+        ALPHA,
+        ALPHANUMERIC,
+        ALPHANUMERICSYMBOL,
+        NUMERIC,
+        ALPHASPACE
+    }
+
     private var _binding: FragmentGeneratorBinding? = null
     private lateinit var myContext: Context
 
@@ -42,11 +61,6 @@ class GeneratorFragment : Fragment() {
 
     private var selectedType = 0
     private var selectedPasswordType = PasswordTypes.ALPHANUMERIC
-
-    companion object {
-        const val TYPE_PASSPHRASE = 1
-        const val TYPE_PASSWORD = 2
-    }
 
     override fun onCreateView(
         inflater: LayoutInflater,
@@ -156,7 +170,7 @@ class GeneratorFragment : Fragment() {
 
     private fun generatePassword() {
         val length = binding.slider.value
-        binding.output.text = generatorPassword(selectedPasswordType, length.toInt())
+        binding.output.text = generatePassword(selectedPasswordType, length.toInt())
     }
 
     private fun storeToClipboard(text: String) {
@@ -172,6 +186,31 @@ class GeneratorFragment : Fragment() {
 
     private fun updateSliderLengthText(length: Int) {
         binding.sliderText.text = getString(R.string.length, ": $length")
+    }
+
+    private fun generatePassword(passwordType: PasswordTypes, length: Int): String {
+        val stringBuilder = StringBuilder()
+        val characters = when (passwordType) {
+            PasswordTypes.ALPHA -> alpha
+            PasswordTypes.ALPHANUMERIC -> alphaNumeric
+            PasswordTypes.ALPHANUMERICSYMBOL -> alphaNumericSymbols
+            PasswordTypes.NUMERIC -> numeric
+            PasswordTypes.ALPHASPACE -> alphaSpace
+        }
+
+        val characterLength = characters.length
+        val secureRandom = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+            SecureRandom.getInstanceStrong()
+        } else {
+            SecureRandom()
+        }
+
+        repeat(length) {
+            val index = secureRandom.nextInt(characterLength)
+            stringBuilder.append(characters[index])
+        }
+
+        return stringBuilder.toString()
     }
 
     override fun onDestroyView() {
